requirements.txt
# Core dependencies
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0

# Web framework for APIs
flask>=2.0.0
requests>=2.28.0

# Database and storage (optional)
boto3>=1.26.0
redis>=4.0.0

# MQTT messaging (optional)
paho-mqtt>=1.6.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Monitoring and metrics
prometheus-client>=0.15.0

# Configuration and utilities
python-dotenv>=0.19.0
pyyaml>=6.0